+++ redbot/cogs/mutes/abc.py
@@ -1,5 +1,5 @@
 from abc import ABC, abstractmethod
-from typing import List, Tuple, Optional, Dict
+from typing import List, Tuple, Optional, Dict, Union
 from datetime import datetime
 
 import discord
@@ -24,4 +24,16 @@ class MixinMeta(ABC):
     async def _voice_perm_check(
         ctx: commands.Context, user_voice_state: Optional[discord.VoiceState], **perms: bool
     ) -> bool:
+        raise NotImplementedError()
+
+    @abstractmethod
+    async def _send_dm_notification(
+        self,
+        user: Union[discord.User, discord.Member],
+        moderator: Union[discord.User, discord.Member],
+        guild: discord.Guild,
+        mute_type: str,
+        reason: str,
+        duration=None,
+    ):
         raise NotImplementedError()
